Use AUG_CCITT instead of unspecified CCITT.
This is to support language-agnosticism: most crc16 implementors
seem to take their table of parameters from the following page:
http://reveng.sourceforge.net/crc-catalogue/16.htm

Bare CCITT parameters are not on that list, so bare CCITT implementations
are not easy to find. For simple replication of results cross-platform,
we use AUG_CCITT instead.

This necessitated changing out the go library used.