Hack address serialization
When addresses were strings, msgp knew all about how to deal with them.
As structs, however, msgp is a bit fussier: it only cares to deal with
the public fields of a struct. Making the contained string public
would entirely defeat the purpose of using an opaque object.

What we therefore do is make it public, generate the msgpack code,
make it private, and then edit the generated code so it still works.
We then disable code generation for this module, because we want to
keep our edits.

It's a bit hacky, but it appears to be the best way to have msgp
serialize the addresses.