Rates are Decimals
We wrap the Decimal type, because that's the go idiom for a transparent
typedef, for reasons. This lets us call all Decimal methods directly
on Rate. It's not perfect: Decimal methods expect pointers, and msgp
codegen expects the type to contain no pointers, so usage could be better.
At least everything appears to work for now, though.