Add soundness test which calculates EAI manually; fix EAI calc
My math-spidey-sense had been tingling after that last commit, so
it was time to take things a bit out of the abstract and just
run a calculation to see if we got the right answer.

We didn't.

Luckily, the problem proved to be tractable; eai.Calculate got
much simpler, and the code reads a bit better now.