Apply a 128-bit context to default rate tables.
Rate serialization doesn't include information about the context,
so we have to manually standardize on one. We use Context128
throughout this project.

However, because rate tables were previously serialized with a default
context, it turned out that there were bit differences when deserialized
with a Context128. Setting the appropriate context when creating the rates
in the first place means that we get back the same numbers we put in
initialiy in downstream tests.