Implement taylor series ExpFrac for unsigned numbers
Generics sure would have been nice here; it feels pretty silly that
we're manually monomorphizing implementations of a bunch of math
across int64/uint64, particularly when the move from signed to
unsigned implementation is literally a find and replace operation
and removing the two test cases on negative numbers. Oh well, it's
go; what can you do.

Also replace the variables a and b with numerator and denominator
in both implementations, because IMO descriptivity is more important
than concision when naming variables.