remove serialization-related code from pkg/key
For consistency, we _must_ use the same serialization and
deserialization code everywhere throughout the codebase.

We have generic code which handles this in pkg/signature:
PublicKey and PrivateKey implement io.Text(Un)Marshaller.

In the future we will update those methods to produce values
which look like the values previously generated.