Update JS API: sign with our signature pkg, so we can read the result
Fundamentally, this just implies adding a type byte to the bytes of the
signature. However, this also changes the data being signed:
signature.PrivateKey.Sign() follows btcec recommendations and
pre-hashes the data to be signed, and signs the hash; the previous
implementation signed the data directly. This means that the actual
signature bytes change somewhat dramatically.