Add Key interface, hide keyBase type
This substantially improves the ergonomics of working with keys
read from user input: the prefixes made it obvious which was correct
when serialized, but that information was lost when cast to the keyBase
type (formerly known as Key).

Making Key into an interface preserves the type information while
retaining flexibility.