Revert "delegate to calculateEAIFactor in CalculateEAIRate"
This reverts commit 41f61cb90ad60ce4cf5aac2df983d4d6c1b68476.

The result from CalculateEAIRate is expected not to be the total
composite rate over the specified period, but rather the annualized
instantaneous rate at a given moment for the specified scenario.

This isn't much use in EAI calculations, but it's easier for humans
to understand. However, it also means that we need to go in and
manually adjust the WAA here in order to get the results we expect.