remove decimal package for go module compatibility
This is actually a fairly vicious change, because much of our test
code depended on that library. Where there was good documentation,
we preserved it; otherwise, we just removed the files.

In the future, we should look into using a different big number
package to restore the test cases. If we find a good candidate,
the first commit in the branch implementing the change should
be to revert this commit. In the meantime, it's even more unsafe to
modify the eai package than usual.